Remove vanish from deopped players
Removing vanish for online deops is flawless, but the implementation of removing
vanish for offline deops leaves something to be desired as we merely refuse to
load the vanish state from disk if the player's permission level is not enough
to enter vanish. While this likely ensures compatibility with mods that change
the way player data is stored, it could break in conditions where the player's
permission level hasn't yet been set during data load (but it seems like with no
other mods present, that won't happen) and will make vanish enabled on a non-op
by an operator reset after rejoin.